requirements.txt
selenium~=4.18.1
pytest~=8.0.1
Appium-Python-Client~=3.1.1
openpyxl~=3.1.2
requests~=2.31.0